Coroutines + Firebase 🔥
This one is a decent refactor, but I put the firestore behind an interface, and then wrapped it with a repository.

The repository exposes suspend functions so that the ViewModel can not have a bunch of Firebase query code, and also we can
concurrently update the database + update state. Now updates are reflected immediately so no more mini delay.